fix issue with id type

+++ src/main/java/mate/academy/service/MovieService.java
@@ -5,5 +5,5 @@ import mate.academy.model.Movie;
 public interface MovieService {
     Movie add(Movie movie);
 
-    Movie get(int id);
+    Movie get(Long id);
 }
